+++ package.json
@@ -11,9 +11,10 @@
   "license": "GPL v3",
   "scripts": {
     "build": "tsc",
-
+    "clean": "rimraf ./dist",
     "lint": "tslint -p ./",
-
+    "poc-ts": "ts-node --project poc/ts/tsconfig.json poc/ts/index.ts",
+    "prebuild": "npm-run-all clean",
     "test": "echo \"Ya we probably need to have some unit tests...\""
   },
   "dependencies": {
@@ -22,7 +23,9 @@
   "devDependencies": {
     "@types/node": "^10.5.1",
     "@types/usb": "^1.1.6",
+    "npm-run-all": "^4.1.3",
     "rimraf": "^2.6.2",
+    "ts-node": "^7.0.0",
     "tslint": "^5.10.0",
     "typescript": "^2.9.2"
   }
