+++ src/models/packet.ts
@@ -1,0 +1,4 @@
+export abstract class Packet {
+
+  public abstract buildPacketBytes(): number[];
+}
