+++ tslint.json
@@ -3,19 +3,24 @@
     "extends": [
         "tslint:recommended"
     ],
+    "linterOptions": {
+        "exclude": [
+            "POC"
+        ]
+    },
     "jsRules": {},
-  "rules": {
-      "trailing-comma": [
-          false
-      ],
-      "max-classes-per-file": [
-          false
-      ],
-      "max-line-length": [
-          false
-      ],
-      "no-trailing-whitespace": false,
-      "eofline": false
-  },
-  "rulesDirectory": []
+    "rules": {
+        "trailing-comma": [
+            false
+        ],
+        "max-classes-per-file": [
+            false
+        ],
+        "max-line-length": [
+            false
+        ],
+        "no-trailing-whitespace": false,
+        "eofline": false
+    },
+    "rulesDirectory": [],
 }
